General update 05/10/22

+++ src/com/jcoadyschaebitz/neon/entity/Item/HealthKit.java
@@ -46,7 +46,7 @@ public class HealthKit extends Item {
 	}
 
 	public void update() {
-		if (level.isPlayerInRad(x, y + yROffset, 15)) doPickUp();
+		if (level.isPlayerInRad((int)x, (int) (y + yROffset), 15)) doPickUp();
 	}
 	
 	@Override
